pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sage"
version = "0.1.0"
description = "RAG-powered product recommendation system"
requires-python = ">=3.11"
dependencies = [
    "huggingface-hub>=0.20.0",
    "sentence-transformers>=2.2.0",
    "transformers>=4.36.0",
    "torch>=2.0.0",
    "qdrant-client>=1.7.0",
    "numpy>=1.24.0",
    "safetensors>=0.4.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.18.0"]
openai = ["openai>=1.12.0"]
pipeline = [
    "anthropic>=0.18.0",
    "datasets>=2.14.0",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
    "matplotlib>=3.7.0",
    "ragas>=0.1.0",
    "requests>=2.28.0",
]
dev = ["pytest>=7.0.0", "pytest-asyncio>=0.21.0", "httpx>=0.25.0"]
api = [
    "anthropic>=0.18.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "prometheus-client>=0.20.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["sage*"]
